Refactor web page tests, fix broken test
Also remove unnecessary module level setup()s
